add concurrentMap method; using ConcurrentSkipListMap for cache
comment: ConcurrentSkipListMap allows modification of map, while traversing it's entries, without ConcurrentModificationException

This is useful when user closes mainWindow while search is in process (during search cache is updated, but closing mainWindows means saving cache entries). It is possible that in the end not all cached entries are saved in their actual state, but it's not critical. Important is that the entries that were already in cache are not lost